+++ designsafe/apps/api/notifications/receivers.py
@@ -15,6 +15,7 @@ WEBSOCKETS_FACILITY = 'websockets'
 @receiver(post_save, sender=Notification, dispatch_uid='notification_msg')
 def send_notification_ws(sender, instance, created, **kwargs):
     #Only send WS message if it's a new notification not if we're updating.
+    logger.debug('receiver received something.')
     if not created:
         return
     try: