+++ designsafe/apps/applications/static/designsafe/apps/applications/scripts/controllers/application-add.js
@@ -67,7 +67,7 @@
                 "type": "string",
                 "description": "Categorization for this app if made public",
                 "enum": [
-                    "Simulation", "Visualization", "Data Processing", "Utilities"
+                    "Simulation", "Visualization", "Data Processing", "Data Collections", "Utilities"
                 ],
                 "title": "Category"
             },
@@ -142,9 +142,6 @@
                 required: function (value) {
                     return value ? true : false;
                 }
-            },
-            validationMessage: {
-                'required': 'Missing required'
             }
         },
         "isPublic",
@@ -211,7 +208,7 @@
                   "type": "string",
                   "description": "Categorization for this app if made public",
                   "enum": [
-                      "Simulation", "Visualization", "Data Processing", "Utilities"
+                      "Simulation", "Visualization", "Data Processing", "Data Collections", "Utilities"
                   ],
                   "title": "Category"
               },
@@ -681,9 +678,6 @@
                               required: function (value) {
                                   return value ? true : false;
                               }
-                          },
-                          validationMessage: {
-                              'required': 'Missing required'
                           }
                       },
                       "shortDescription",
@@ -1589,6 +1583,13 @@
             switch($scope.addModel.select){
               case 'Agave':
                 if ($scope.myForm.$valid){
+                    // Add formatted appCategory entry to tags, and replace old appCategory if it exists
+                    if ($scope.model.hasOwnProperty('appCategory')) {
+                        if ($scope.model.tags.filter(s => s.includes('appCategory')) !== undefined && $scope.model.tags.filter(s => s.includes('appCategory')).length != 0) {
+                            $scope.model.tags.splice($scope.model.tags.indexOf($scope.model.tags.filter(s => s.includes('appCategory'))[0]));
+                        }
+                        $scope.model.tags.push(`appCategory:${$scope.model.appCategory}`);
+                    }
                   Apps.createApp($scope.model)
                     .then(
                       function(response){
@@ -1618,7 +1619,10 @@
                                         },
                                         controller: [
                                          '$scope', '$uibModalInstance', '$translate', 'appMeta', function($scope, $uibModalInstance, $translate, appMeta) {
-
+                                            // Define appCategory if it exists in tags
+                                            if (appMeta.tags.filter(s => s.includes('appCategory')) !== undefined && appMeta.tags.filter(s => s.includes('appCategory')).length != 0) {
+                                                appMeta.appCategory = appMeta.tags.filter(s => s.includes('appCategory'))[0].split(':')[1];
+                                            }
                                             $scope.appMeta = appMeta;
 
                                             $scope.close = function() {
@@ -1651,6 +1655,10 @@
                                         },
                                         controller: [
                                          '$scope', '$uibModalInstance', '$translate', 'appMeta', function($scope, $uibModalInstance, $translate, appMeta) {
+                                            // Define appCategory if it exists in tags
+                                            if (appMeta.definition.tags.filter(s => s.includes('appCategory')) !== undefined && appMeta.definition.tags.filter(s => s.includes('appCategory')).length != 0) {
+                                                 appMeta.definition.appCategory = appMeta.definition.tags.filter(s => s.includes('appCategory'))[0].split(':')[1];
+                                            }
                                             $scope.appMeta = appMeta;
                                             $scope.close = function() {
                                               $uibModalInstance.dismiss();
@@ -1693,6 +1701,7 @@
                   metadata.value.definition.id = $scope.customModel.label+ '-' + $scope.customModel.version;
                   metadata.value.definition.available = true;
                   metadata.value.definition.isPublic = false;
+                  metadata.value.definition.appCategory = $scope.customModel.appCategory;
                   metadata.value.type = 'html';
                   _.extend(metadata.value.definition, angular.copy($scope.customModel));