+++ designsafe/apps/workspace/static/designsafe/apps/workspace/scripts/services/jobs-service.js
@@ -22,6 +22,11 @@
       return $http.post(djangoUrl.reverse('designsafe_workspace:call_api', ['jobs']), data);
     };
 
+    service.getWebhookUrl = function() {
+      // console.log(djangoUrl.reverse('designsafe_api:jobs_wb_handler'))
+      return $http.get(djangoUrl.reverse('designsafe_api:jobs_wh_handler'));  
+    }
+
     return service;
   }]);
 })(window, angular, jQuery, _);