+++ designsafe/middleware.py
@@ -16,6 +16,22 @@ from termsandconditions.models import TermsAndConditions
 
 logger = logging.getLogger(__name__)
 
+class DesignSafeSupportedBrowserMiddleware:
+    """
+    Middleware to check if the user is running Chrome or Firefox and 
+    flash a warning otherwise.
+    """
+
+    def process_request(self, request):
+        user_agent = request.META['HTTP_USER_AGENT']
+        agent_is_supported = ('Chrome' in user_agent) or ('Firefox' in user_agent)
+
+        if not agent_is_supported:
+            messages.warning(request, '<h4>Unsupported Browser</h4>'
+                                      'Your browser is not supported by DesignSafe. '
+                                      'Please switch to <a href="https://www.google.com/chrome">Chrome</a> '
+                                      'or <a href="https://www.mozilla.org/en-US/firefox/new/">Firefox</a> '
+                                      'if you experience issues.')
 
 class DesignSafeTermsMiddleware(TermsAndConditionsRedirectMiddleware):
     """