+++ designsafe/static/scripts/data-depot/controllers/my-data.js
@@ -35,7 +35,7 @@
     $scope.onBrowse = function ($event, file) {
       $event.preventDefault();
       $event.stopPropagation();
-      if (file.type === 'file') {
+      if (typeof(file.type) !== 'undefined' && file.type !== 'dir' && file.type !== 'folder') {
         DataBrowserService.preview(file, $scope.browser.listing);
       } else {
         $state.go('myData', {systemId: file.system, filePath: file.path});