+++ designsafe/static/scripts/data-depot/controllers/projects.js
@@ -285,6 +285,187 @@
         });
     }
 
+    function checkMetadata(){
+      var projData = $scope.data.project.value;
+      var experimentsList = $scope.state.publication.experimentsList;
+      var analysisList = $scope.state.publication.analysisList;
+      var reportsList = $scope.state.publication.reportsList;
+      var publicationMessages = [];
+      var checklist = {
+        project: {},
+      };
+
+      // object containing required* fields that will be checked on metadata page
+      // TODO:  add requirements to simulation and other project types
+      var requirements = {
+        "projectReq": ['title', 'projectType', 'teamMembers', 'description', 'awardNumber', 'keywords'],
+        "experimentReq": ['title', 'experimentType', 'experimentalFacility', 'description', 'authors'],
+        "simulationReq": [],
+        "otherReq": [],
+        "reportReq": ['title', 'description'],
+        "analysisReq": ['title', 'description'],
+      };
+
+      //TODO: set 'requirement checks' in their own functions
+      //check project requirements
+      function projectRequirements(){
+        checklist.project.name = projData.title;
+        checklist.project.category = "project";
+        requirements.projectReq.forEach(function(req) {
+          if (projData[req] == '' || projData[req] == []) {
+            checklist.project[req] = false;
+          } else {
+            checklist.project[req] = true;
+          }
+        });
+        return;
+      }
+      
+      //check experiment requirements
+      function experimentRequirements(){
+        var i = 0;
+        //TODO: remove these with revision of event,sensor,modelconf check
+        var models = $scope.data.project.modelconfig_set;
+        var sensors = $scope.data.project.sensorlist_set;
+        var events = $scope.data.project.event_set;
+
+        experimentsList.forEach(function (exp) {
+          var title = experimentsList[i].value.title;
+
+          checklist['experiment'+i] = {};
+          checklist['experiment'+i].name = title;
+          checklist['experiment'+i].category = 'experiment';
+
+          
+          //model configs
+          checklist['experiment'+i].model = false;
+          models.forEach(function(mod){
+            mod.value.experiments.forEach(function(mid) {
+              if (exp.uuid == mid) {
+                checklist['experiment'+i].model = true;
+              }
+            });
+          });
+          
+          //sensor data
+          checklist['experiment'+i].sensor = false;
+          sensors.forEach(function(sen) {
+            sen.value.experiments.forEach(function(sid) {
+              if (exp.uuid == sid) {
+                checklist['experiment'+i].sensor = true;
+              }
+            });
+          });
+
+          //event data
+          checklist['experiment'+i].event = false;
+          events.forEach(function(evt) {
+            evt.value.experiments.forEach(function(eid) {
+              if (exp.uuid == eid) {
+                checklist['experiment'+i].event = true;
+              }
+            });
+          });
+
+          requirements.experimentReq.forEach(function (req) {
+            if (exp.value[req] == '' || exp.value[req] == []) {
+              checklist['experiment'+i][req] = false;
+            } else {
+              checklist['experiment'+i][req] = true;
+            }
+          });
+          i++;
+        });
+        return;
+      }
+
+      //check analysis requirements
+      function analysisRequirements(){
+        var i = 0;
+        analysisList.forEach(function (exp) {
+          var title = analysisList[i].value.title;
+
+          checklist['analysis'+i] = {};
+          checklist['analysis'+i].name = title;
+          checklist['analysis'+i].category = 'analysis';
+          requirements.analysisReq.forEach(function (req) {
+            if (exp.value[req] == '' || exp.value[req] == []) {
+              checklist['analysis'+i][req] = false;
+            } else {
+              checklist['analysis'+i][req] = true;
+            }
+          });
+          i++;
+        });
+        return;
+      }
+
+      //check report requirements
+      function reportRequirements(){
+        var i = 0;
+        reportsList.forEach(function (exp) {
+          var title = reportsList[i].value.title;
+
+          checklist['report'+i] = {};
+          checklist['report'+i].name = title;
+          checklist['report'+i].category = 'report';
+          requirements.reportReq.forEach(function (req) {
+            if (exp.value[req] == '' || exp.value[req] == []) {
+              checklist['report'+i][req] = false;
+            } else {
+              checklist['report'+i][req] = true;
+            }
+          });
+          i++;
+        });
+        return;
+      }
+
+      
+      projectRequirements();
+      experimentRequirements();
+      analysisRequirements();
+      reportRequirements();
+
+      //TODO:   tune this area to respond to project type before checking
+      //check for missing metadata (experimental)
+      var allow = true;
+      if (projData.projectType == "experimental") {
+        // check for missing records
+        if (Object.keys(checklist).includes('experiment0') !== true) {
+          publicationMessages.push({title: "Project", message: "Your project must include an experiment."});
+          allow = false;
+        }
+      }
+
+      i = 0;
+      Object.values(checklist).forEach(function(exp) {
+        Object.entries(exp).forEach(function(res) {
+          // res[0] == keys
+          // res[1] == values
+          if (res[1] === false) {
+            if (res[0] == 'title') {
+              //if 'title' is missing push data category type
+              publicationMessages.push({title: exp.category, message: "You are missing: " + res[0]});
+              allow = false;
+            } else {
+              //if anything else push the data name
+              publicationMessages.push({title: exp.name, message: "You are missing: " + res[0]});
+              allow = false;
+            }
+          }
+        });
+        i++;
+      });
+
+      if (allow) {
+        $scope.state.publishPipeline = 'agreement';
+      } else {
+        $scope.ui.publicationMessages = publicationMessages;
+        return;
+      }
+    }
+
     $scope.publishPipeline_start = function(){
       $scope.state.publishPipeline = 'select';
     };
@@ -348,7 +529,7 @@
         $scope.state.publishPipeline = 'meta';
       }
       else if (st == 'meta'){
-        $scope.state.publishPipeline = 'agreement';
+        checkMetadata();
       }else {
         $scope.state.publishPipeline = 'agreement';
       }
@@ -372,6 +553,11 @@
       var modelConfigs = [];
       var sensorLists = [];
       var publicationMessages = [];
+
+      // TODO:  A series of checks and notification flags were created here 
+      //        but they are not being displayed in the metadata section.
+      //        Suggest going over these since some will be redundant. See
+      //        above "checkMetadata"
       
       if ($scope.state.project.value.projectType == 'experimental'){
         if (publication.experimentsList){
@@ -751,7 +937,7 @@
     $scope.onBrowseData = function onBrowseData($event, file) {
       $event.preventDefault();
       DataBrowserService.showListing();
-      if (file.type === 'file') {
+      if (typeof(file.type) !== 'undefined' && file.type !== 'dir' && file.type !== 'folder') {
         DataBrowserService.preview(file, $scope.browser.listing);
       } else {
         $state.go('projects.view.data', {projectId: projectId,
@@ -1037,6 +1223,21 @@
 
       openEditCategories: function(){
         DataBrowserService.viewCategories([]);
+      },
+
+      isMissing: function(ent){
+        if (typeof(ent) == 'object') {
+          if (Object.keys(ent).length < 1) {
+            return true;
+          } else {
+            return false;
+        }
+        }
+        if (ent === '' || ent === undefined) {
+          return true;
+        } else {
+          return false;
+        }
       },
 
       selectAllFiles : function(ent, evt){